Implmentaçãoes importantes com o Riskengine->FixFixSender

+++ oms/ServiceDefaults/AccountRiskState.cs
@@ -9,8 +9,8 @@ namespace ServiceDefaults;
 public struct AccountRiskState
 {
     [FieldOffset(0)] public long AccountId;
-    [FieldOffset(8)] public double AvailableCash;         // Saldo D-0
-    [FieldOffset(16)] public double BlockedCash;           // Margem de ordens abertas
+    [FieldOffset(8)] public decimal AvailableCash;         // Saldo D-0
+    [FieldOffset(16)] public decimal BlockedCash;           // Margem de ordens abertas
     [FieldOffset(24)] public int MaxOrderQuantity;         // Fat-finger limite
     [FieldOffset(28)] public int TotalTradedQuantityToday;
     [FieldOffset(32)] public bool IsBlocked;               // Status da conta (true = kill switch)
